Remove root/cube/square functions
