+++ NAMESPACE
@@ -305,12 +305,10 @@ export(rray_axis_names)
 export(rray_bind)
 export(rray_broadcast)
 export(rray_cbind)
-export(rray_cbrt)
 export(rray_clip)
 export(rray_col_names)
 export(rray_cos)
 export(rray_cospi)
-export(rray_cube)
 export(rray_diag)
 export(rray_dim)
 export(rray_dim_common)
@@ -369,8 +367,6 @@ export(rray_slice)
 export(rray_slice_assign)
 export(rray_sort)
 export(rray_split)
-export(rray_sqrt)
-export(rray_square)
 export(rray_squeeze)
 export(rray_subset)
 export(rray_subset_assign)
