--- R/math-power.R
@@ -1,60 +1,0 @@
-#' Power functions
-#'
-#' @description
-#'
-#' - `rray_square()` - computes the elementwise square of `x` (i.e. `x * x`).
-#'
-#' - `rray_cube()` - computes the elementwise cube of `x` (i.e. `x * x * x`).
-#'
-#' - `rray_sqrt()` - computes the elementwise square root of `x`.
-#'
-#' - `rray_cbrt()` - computes the elementwise cube root of `x`.
-#'
-#' - `rray_hypot()` - computes the elementwise square root of the sum
-#' of squares of `x` and `y`.
-#'
-#' @param x,y A vector, matrix, array or rray.
-#'
-#' @examples
-#' x <- matrix(c(2, 4, 6))
-#'
-#' rray_square(x)
-#' rray_sqrt(x)
-#'
-#' rray_cube(x)
-#' rray_cbrt(x)
-#'
-#' # With broadcasting
-#' rray_hypot(x, t(x))
-#'
-#' @family power math functions
-#' @export
-rray_square <- function(x) {
-  vec_cast_container(rray__square(x), x)
-}
-
-#' @rdname rray_square
-#' @export
-rray_cube <- function(x) {
-  vec_cast_container(rray__cube(x), x)
-}
-
-#' @rdname rray_square
-#' @export
-rray_sqrt <- function(x) {
-  vec_cast_container(rray__sqrt(x), x)
-}
-
-#' @rdname rray_square
-#' @export
-rray_cbrt <- function(x) {
-  vec_cast_container(rray__cbrt(x), x)
-}
-
-#' @rdname rray_square
-#' @export
-rray_hypot <- function(x, y) {
-  out <- rray__hypot(x, y)
-  container <- vec_type_container2(x, y)
-  vec_cast_container(out, container)
-}
