+++ man/rray_hypot.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/hypot.R
+\name{rray_hypot}
+\alias{rray_hypot}
+\title{Compute the square root of the sum of squares}
+\usage{
+rray_hypot(x, y)
+}
+\arguments{
+\item{x, y}{A vector, matrix, array or rray.}
+}
+\description{
+\code{rray_hypot()} computes the elementwise square root of the sum
+of squares of \code{x} and \code{y}.
+}
+\examples{
+x <- matrix(c(2, 4, 6))
+
+# With broadcasting
+rray_hypot(x, t(x))
+
+}
