--- man/rray_square.Rd
@@ -1,47 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/math-power.R
-\name{rray_square}
-\alias{rray_square}
-\alias{rray_cube}
-\alias{rray_sqrt}
-\alias{rray_cbrt}
-\alias{rray_hypot}
-\title{Power functions}
-\usage{
-rray_square(x)
-
-rray_cube(x)
-
-rray_sqrt(x)
-
-rray_cbrt(x)
-
-rray_hypot(x, y)
-}
-\arguments{
-\item{x, y}{A vector, matrix, array or rray.}
-}
-\description{
-\itemize{
-\item \code{rray_square()} - computes the elementwise square of \code{x} (i.e. \code{x * x}).
-\item \code{rray_cube()} - computes the elementwise cube of \code{x} (i.e. \code{x * x * x}).
-\item \code{rray_sqrt()} - computes the elementwise square root of \code{x}.
-\item \code{rray_cbrt()} - computes the elementwise cube root of \code{x}.
-\item \code{rray_hypot()} - computes the elementwise square root of the sum
-of squares of \code{x} and \code{y}.
-}
-}
-\examples{
-x <- matrix(c(2, 4, 6))
-
-rray_square(x)
-rray_sqrt(x)
-
-rray_cube(x)
-rray_cbrt(x)
-
-# With broadcasting
-rray_hypot(x, t(x))
-
-}
-\concept{power math functions}
